# api/openapi.yaml
openapi: 3.0.3
info:
  title: Shop Public API
  version: 1.4.0
  description: |
    Public and back-office API of the demo shop platform: catalog, carts, orders,
    payments, shipments, reviews, promo codes and the user back office.

    The document is maintained by hand and is the contract of record for the
    `shop-api-spring` service. It is intentionally NOT generated from the running
    application.
  contact:
    name: Shop Platform Team
    email: platform@shop.example.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/shop
    description: Local developer instance (server.servlet.context-path = /shop)
  - url: https://api.shop.example.com/shop
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/orders:
    get:
      operationId: listOrders
      summary: List orders
      tags: [orders]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: customerEmail
          in: query
          required: false
          schema:
            type: string
            format: email
            maxLength: 160
      responses:
        '200':
          description: Page of orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createOrder
      summary: Place a new order
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Order created
          headers:
            Location:
              description: URI of the created order
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/summary:
    get:
      operationId: getOrderSummary
      summary: Aggregated order statistics for a date range
      description: 'Requires the `X-Api-Version: 2` request header; version 1 clients are not routed here.'
      tags: [orders]
      parameters:
        - name: X-Api-Version
          in: header
          required: true
          description: Contract version selector, must be `2`
          schema:
            type: string
            enum: ['2']
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Order summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{id}:
    parameters:
      - $ref: '#/components/parameters/OrderIdParam'
    get:
      operationId: getOrder
      summary: Get a single order
      tags: [orders]
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteOrder
      summary: Delete an order
      tags: [orders]
      responses:
        '204':
          description: Order deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{id}/status:
    put:
      operationId: updateOrderStatus
      summary: Move an order to a new status
      tags: [orders]
      parameters:
        - $ref: '#/components/parameters/OrderIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusUpdateRequest'
      responses:
        '200':
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{id}/export:
    get:
      operationId: exportOrderInvoice
      summary: Download the order invoice as PDF
      tags: [orders]
      parameters:
        - $ref: '#/components/parameters/OrderIdParam'
        - name: action
          in: query
          required: true
          description: Discriminating query parameter, must be `export`
          schema:
            type: string
            enum: [export]
      responses:
        '200':
          description: Invoice document
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/carts/{id}:
    parameters:
      - $ref: '#/components/parameters/CartIdParam'
    get:
      operationId: getCart
      summary: Get a shopping cart
      tags: [carts]
      security: []
      responses:
        '200':
          description: Cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: patchCart
      summary: Update cart owner or promo code
      tags: [carts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartPatchRequest'
      responses:
        '200':
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/carts/{id}/items:
    post:
      operationId: addCartItem
      summary: Add an item to the cart
      tags: [carts]
      security: []
      parameters:
        - $ref: '#/components/parameters/CartIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemRequest'
      responses:
        '200':
          description: Cart with the item added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/carts/{id}/items/{sku}:
    delete:
      operationId: removeCartItem
      summary: Remove an item from the cart
      tags: [carts]
      security: []
      parameters:
        - $ref: '#/components/parameters/CartIdParam'
        - name: sku
          in: path
          required: true
          description: Stock keeping unit of the item to remove
          schema:
            type: string
            pattern: '^SKU-[0-9]{5}$'
      responses:
        '200':
          description: Cart without the removed item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /admin/users:
    get:
      operationId: listUsers
      summary: List back-office and customer accounts
      tags: [admin]
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: role
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UserRole'
        - name: blocked
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Page of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/users/{id}:
    get:
      operationId: getUser
      summary: Get a single account
      tags: [admin]
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/users/{id}/role:
    put:
      operationId: changeUserRole
      summary: Change the role of an account
      tags: [admin]
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - name: X-Admin-Actor
          in: header
          required: false
          description: Login of the operator performing the change, recorded in the audit trail
          schema:
            type: string
            maxLength: 80
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleUpdateRequest'
      responses:
        '200':
          description: Role changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Customer or operator access token issued by the shop identity provider
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Static back-office key, required for every /admin operation
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      description: Zero based page index
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      description: Page size
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 20
...
    OrderIdParam:
      name: id
      in: path
      required: true
      description: Order identifier
      schema:
        type: string
        maxLength: 40
...
    CartIdParam:
      name: id
      in: path
      required: true
      description: Cart identifier
      schema:
        type: string
        maxLength: 40
...
    UserIdParam:
      name: id
      in: path
      required: true
      description: User identifier
      schema:
        type: string
        maxLength: 40
  responses:
    BadRequest:
      description: The request payload or parameters failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The caller is authenticated but not allowed to perform the operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: The operation conflicts with the current state of the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unexpected server failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
          example: VALIDATION_FAILED
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          items:
            type: string
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: double
          minimum: 0
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          example: RUB
    Address:
      type: object
      required: [country, city, street, postalCode]
      properties:
        country:
          type: string
          pattern: '^[A-Z]{2}$'
        city:
          type: string
          maxLength: 80
        street:
          type: string
          maxLength: 160
        postalCode:
          type: string
          pattern: '^[0-9]{6}$'
...
    OrderStatus:
      type: string
      enum: [NEW, PAID, PACKING, SHIPPED, DELIVERED, CANCELLED]
    OrderItem:
      type: object
      required: [sku, title, quantity, unitPrice]
      properties:
        sku:
          type: string
          pattern: '^SKU-[0-9]{5}$'
        title:
          type: string
        quantity:
          type: integer
          format: int32
          minimum: 1
          maximum: 999
        unitPrice:
          $ref: '#/components/schemas/Money'
    Order:
      type: object
      required: [id, customerEmail, status, items, total]
      properties:
        id:
          type: string
        customerEmail:
          type: string
          format: email
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        total:
          $ref: '#/components/schemas/Money'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        promoCode:
          type: string
          pattern: '^[A-Z0-9]{4,16}$'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderCreateRequest:
      type: object
      required: [customerEmail, items, shippingAddress]
      properties:
        customerEmail:
          type: string
          format: email
          maxLength: 160
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItem'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        promoCode:
          type: string
          pattern: '^[A-Z0-9]{4,16}$'
    OrderStatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'
        comment:
          type: string
          maxLength: 300
    OrderSummary:
      type: object
      required: [ordersCount, revenue]
      properties:
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        ordersCount:
          type: integer
          format: int64
        revenue:
          $ref: '#/components/schemas/Money'
        byStatus:
          type: object
          description: Order counters keyed by order status
          additionalProperties:
            type: integer
            format: int64
    OrderPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Order'
    CartItem:
      type: object
      required: [sku, quantity, unitPrice]
      properties:
        sku:
          type: string
          pattern: '^SKU-[0-9]{5}$'
        title:
          type: string
        quantity:
          type: integer
          format: int32
          minimum: 1
          maximum: 99
        unitPrice:
          $ref: '#/components/schemas/Money'
    Cart:
      type: object
      required: [id, items, subtotal]
      properties:
        id:
          type: string
        customerEmail:
          type: string
          format: email
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        subtotal:
          $ref: '#/components/schemas/Money'
        promoCode:
          type: string
          pattern: '^[A-Z0-9]{4,16}$'
        updatedAt:
          type: string
          format: date-time
    CartItemRequest:
      type: object
      required: [sku, quantity]
      properties:
        sku:
          type: string
          pattern: '^SKU-[0-9]{5}$'
        quantity:
          type: integer
          format: int32
          minimum: 1
          maximum: 99
    CartPatchRequest:
      type: object
      properties:
        customerEmail:
          type: string
          format: email
          maxLength: 160
        promoCode:
          type: string
          pattern: '^[A-Z0-9]{4,16}$'
...
    UserRole:
      type: string
      enum: [CUSTOMER, SUPPORT, MANAGER, ADMIN]
    UserDto:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
          maxLength: 120
        role:
          $ref: '#/components/schemas/UserRole'
        blocked:
          type: boolean
        registeredAt:
          type: string
          format: date-time
    UserRoleUpdateRequest:
      type: object
      required: [role]
      properties:
        role:
          $ref: '#/components/schemas/UserRole'
        justification:
          type: string
          maxLength: 200
    UserPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/UserDto'